Aurora MySQL is not available pre 202, so split it out of tests etc
