+++ jetbrains-ultimate/tst-201+/software/aws/toolkits/jetbrains/services/rds/actions/CreateConfigurationActionTest.kt
@@ -4,7 +4,6 @@
 package software.aws.toolkits.jetbrains.services.rds.actions
 
 import com.intellij.database.autoconfig.DataSourceRegistry
-import com.intellij.database.remote.jdbc.helpers.JdbcSettings.SslMode
 import com.intellij.testFramework.ProjectRule
 import com.nhaarman.mockitokotlin2.doAnswer
 import com.nhaarman.mockitokotlin2.mock
@@ -23,7 +22,6 @@ import software.aws.toolkits.jetbrains.services.rds.RdsDatasourceConfiguration
 import software.aws.toolkits.jetbrains.services.rds.RdsNode
 import software.aws.toolkits.jetbrains.services.rds.auth.IamAuth
 import software.aws.toolkits.jetbrains.services.rds.jdbcMysql
-import software.aws.toolkits.jetbrains.services.rds.jdbcMysqlAurora
 import software.aws.toolkits.jetbrains.services.rds.jdbcPostgres
 import software.aws.toolkits.jetbrains.services.rds.mysqlEngineType
 import software.aws.toolkits.jetbrains.services.rds.postgresEngineType
@@ -164,46 +162,6 @@ class CreateConfigurationActionTest {
             assertThat(it.driverClass).contains("mysql")
             assertThat(it.url).contains(jdbcMysql)
             assertThat(it.sslCfg).isNull()
-        }
-    }
-
-    @Test
-    fun `Add Aurora MySQL data source`() {
-        val instance = createDbInstance(address = address, port = port, engineType = "aurora")
-        val registry = DataSourceRegistry(projectRule.project)
-        registry.createRdsDatasource(
-            RdsDatasourceConfiguration(
-                username = username,
-                credentialId = MockCredentialsManager.DUMMY_PROVIDER_IDENTIFIER.id,
-                regionId = MockRegionProvider.getInstance().defaultRegion().id,
-                dbInstance = instance
-            )
-        )
-        assertThat(registry.newDataSources).hasOnlyOneElementSatisfying {
-            assertThat(it.username).isEqualTo(username)
-            assertThat(it.driverClass).contains("mariadb")
-            assertThat(it.url).contains(jdbcMysqlAurora)
-            assertThat(it.sslCfg?.myMode).isEqualTo(SslMode.REQUIRE)
-        }
-    }
-
-    @Test
-    fun `Add Aurora MySQL 5_7 data source`() {
-        val instance = createDbInstance(address = address, port = port, engineType = "aurora-mysql")
-        val registry = DataSourceRegistry(projectRule.project)
-        registry.createRdsDatasource(
-            RdsDatasourceConfiguration(
-                username = username,
-                credentialId = MockCredentialsManager.DUMMY_PROVIDER_IDENTIFIER.id,
-                regionId = MockRegionProvider.getInstance().defaultRegion().id,
-                dbInstance = instance
-            )
-        )
-        assertThat(registry.newDataSources).hasOnlyOneElementSatisfying {
-            assertThat(it.username).isEqualTo(username)
-            assertThat(it.driverClass).contains("mariadb")
-            assertThat(it.url).contains(jdbcMysqlAurora)
-            assertThat(it.sslCfg?.myMode).isEqualTo(SslMode.REQUIRE)
         }
     }
